wmap_back_end/requirements.txt
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
httpie==1.0.2
idna==2.7
numpy==1.15.4
pandas==0.23.4
psycopg2==2.7.6.1
Pygments==2.3.0
pyproj==1.9.5.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.0
Shapely==1.6.4.post2
six==1.11.0
urllib3==1.24.1
